SuperSU: Only include if specified - most users prefer Magisk
We can eventually migrate to adding Magisk by default if
people think it's a good idea. Chainfire has not been
keeping his SSL certificates up to date, and is no longer
supporting the SuperSU project. It is now owned by a
corporation with unknown intentions, and we would prefer
not to update the links to point to their site.

Signed-off-by: James Christopher Adduono <jc@adduono.com>