Updates to work with Android Pie devices
Devices shipped with Android 9 must adhere to the new partition layouts:

system as root
A/B

Changes in this commit make the nethunter-installer compatible with the
system as root layout. The A/B partitioning scheme requires further
modifications to the boot-patcher to avoid the recovery image in the
boot partition from being corrupted during flashing.

Signed-off-by: Re4son <3520622-re4son@users.noreply.gitlab.com>